requirements.txt

emoji==0.5.1
lxml==4.3.1
numpy==1.16.1
oauthlib==2.1.0
Pattern==3.6
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.0.0
scikit-learn==0.20.2
scipy==1.2.1
sklearn==0.0
tweepy==3.7.0
tweet-preprocessor==0.5.0
TwitterAPI==2.5.8
Unidecode==1.0.23
urllib3==1.24.1
vaderSentiment==3.2.1

